refactor(search): move semantic query logic
