+++ shared/src/db/repositories/embedding.rs
@@ -1,7 +1,6 @@
 use crate::{
     db::error::DatabaseError,
-    models::{ChunkResult, Document, Embedding},
-    SourceType,
+    models::{Document, Embedding},
 };
 use pgvector::Vector;
 use sqlx::{PgPool, Row};
@@ -14,21 +13,6 @@ pub struct EmbeddingRepository {
 impl EmbeddingRepository {
     pub fn new(pool: &PgPool) -> Self {
         Self { pool: pool.clone() }
-    }
-
-    /// Generate SQL condition to check if user has permission to access document
-    fn generate_permission_filter(&self, user_email: &str) -> String {
-        format!(
-            r#"(
-                (d.permissions->>'public')::boolean = true OR
-                d.permissions->'users' ? '{}' OR
-                d.permissions->'groups' ? ANY(
-                    -- TODO: Add group membership lookup here
-                    ARRAY['{}']::text[]
-                )
-            )"#,
-            user_email, user_email
-        )
     }
 
     pub async fn find_by_document_id(
@@ -123,198 +107,6 @@ impl EmbeddingRepository {
 
         tx.commit().await?;
         Ok(())
-    }
-
-    pub async fn find_similar_with_filters(
-        &self,
-        embedding: Vec<f32>,
-        source_types: Option<&[SourceType]>,
-        content_types: Option<&[String]>,
-        limit: i64,
-        offset: i64,
-        user_email: Option<&str>,
-        document_id: Option<&str>,
-        recency_boost_weight: f32,
-        recency_half_life_days: f32,
-    ) -> Result<Vec<ChunkResult>, DatabaseError> {
-        let dims = embedding.len() as i16;
-        let vector = Vector::from(embedding);
-
-        let mut where_conditions = Vec::new();
-
-        // Filter to matching dimensions so the partial HNSW index is used
-        where_conditions.push(format!("e.dimensions = ${}", 4));
-
-        // Fixed bind slots: $1=vector, $2=limit, $3=offset, $4=dims,
-        // $5=recency_boost_weight, $6=recency_half_life_days.
-        // Dynamic filters (document_id, source_types, content_types) start at $7.
-        let mut bind_index = 7;
-
-        // Filter by the current active embedding model via subquery
-        where_conditions.push(
-            "e.model_name = (SELECT config->>'model' FROM embedding_providers WHERE is_current = TRUE AND is_deleted = FALSE LIMIT 1)"
-                .to_string(),
-        );
-
-        // Add document_id filter if provided
-        if let Some(_) = document_id {
-            where_conditions.push(format!("e.document_id = ${}", bind_index));
-            bind_index += 1;
-        }
-
-        if let Some(src) = source_types {
-            if !src.is_empty() {
-                where_conditions.push(format!(
-                    "d.source_id IN (SELECT id FROM sources WHERE source_type = ANY(${}))",
-                    bind_index
-                ));
-                bind_index += 1;
-            }
-        }
-
-        if let Some(ct) = content_types {
-            if !ct.is_empty() {
-                where_conditions.push(format!("d.content_type = ANY(${})", bind_index));
-            }
-        }
-
-        // Add permission filtering if user email is provided
-        if let Some(email) = user_email {
-            where_conditions.push(self.generate_permission_filter(email));
-        }
-
-        let where_clause = format!("WHERE {}", where_conditions.join(" AND "));
-
-        // Recency-boosted vector search (mirrors the FTS approach in search_repository.rs).
-        //
-        // Strategy: over-fetch 3x candidates by raw cosine distance via the HNSW
-        // index, then re-rank in a second pass that incorporates document age.
-        //
-        // The `<=>` operator returns cosine *distance* in [0, 2] (lower = more similar).
-        // The recency factor is always >= 1.0:
-        //   recency_factor = 1.0 + weight * EXP(-age_seconds / (86400 * half_life_days))
-        //     - Recent docs  → EXP ≈ 1  → factor ≈ 1 + weight  (max boost)
-        //     - Old docs      → EXP ≈ 0  → factor ≈ 1.0         (no boost)
-        //
-        // Dividing distance by recency_factor shrinks the distance for recent docs,
-        // making them rank higher. The caller converts back via `similarity = 1 - distance`,
-        // so the resulting similarity_score already includes the recency boost.
-        //
-        // The effective timestamp prefers `metadata->>'updated_at'` (the original
-        // document date) over `d.updated_at` (which reflects indexing time).
-        let recency_expr = format!(
-            "(1.0 + $5::double precision * EXP(\
-                -EXTRACT(EPOCH FROM (CURRENT_TIMESTAMP - COALESCE(\
-                    CASE WHEN c.doc_metadata->>'updated_at' IS NOT NULL \
-                         AND pg_input_is_valid(c.doc_metadata->>'updated_at', 'timestamptz') \
-                    THEN (c.doc_metadata->>'updated_at')::timestamptz END, \
-                    c.doc_updated_at))) \
-                / (86400.0 * $6::double precision)))::real"
-        );
-
-        let query_str = format!(
-            r#"
-            WITH candidates AS (
-                SELECT
-                    e.document_id,
-                    e.embedding <=> $1 as distance,
-                    e.chunk_start_offset,
-                    e.chunk_end_offset,
-                    e.chunk_index,
-                    d.external_id,
-                    d.updated_at as doc_updated_at,
-                    d.metadata as doc_metadata,
-                    s.source_type
-                FROM embeddings e
-                JOIN documents d ON e.document_id = d.id
-                JOIN sources s ON s.id = d.source_id AND NOT s.is_deleted
-                {where_clause}
-                ORDER BY e.embedding <=> $1
-                LIMIT ($2 + $3) * 3
-            ),
-            scored_candidates AS (
-                SELECT
-                    c.document_id,
-                    c.distance / {recency_expr} as distance,
-                    c.chunk_start_offset,
-                    c.chunk_end_offset,
-                    c.chunk_index,
-                    c.external_id,
-                    c.doc_updated_at,
-                    c.source_type
-                FROM candidates c
-            ),
-            deduped_candidates AS (
-                SELECT document_id, distance, chunk_start_offset, chunk_end_offset, chunk_index
-                FROM (
-                    SELECT sc.*,
-                           ROW_NUMBER() OVER (
-                               PARTITION BY sc.source_type, sc.external_id
-                               ORDER BY sc.distance ASC, sc.doc_updated_at DESC, sc.document_id, sc.chunk_index
-                           ) AS dedupe_rank
-                    FROM scored_candidates sc
-                ) ranked_candidates
-                WHERE dedupe_rank = 1
-            )
-            SELECT
-                dc.document_id,
-                dc.distance,
-                dc.chunk_start_offset,
-                dc.chunk_end_offset,
-                dc.chunk_index
-            FROM deduped_candidates dc
-            ORDER BY distance
-            LIMIT $2 OFFSET $3
-            "#,
-            where_clause = where_clause,
-            recency_expr = recency_expr,
-        );
-
-        let mut query = sqlx::query(&query_str)
-            .bind(&vector)
-            .bind(limit)
-            .bind(offset)
-            .bind(dims)
-            .bind(recency_boost_weight as f64)
-            .bind(recency_half_life_days as f64);
-
-        if let Some(doc_id) = document_id {
-            query = query.bind(doc_id);
-        }
-
-        if let Some(src) = source_types {
-            if !src.is_empty() {
-                query = query.bind(src);
-            }
-        }
-
-        if let Some(ct) = content_types {
-            if !ct.is_empty() {
-                query = query.bind(ct);
-            }
-        }
-
-        let results = query.fetch_all(&self.pool).await?;
-        // Convert boosted cosine distance back to a similarity score.
-        // The distance column already has recency boosting applied in SQL,
-        // so the similarity_score returned here reflects both semantic
-        // relevance and document recency — no further Rust-side boost needed.
-        let chunk_results: Vec<ChunkResult> = results
-            .into_iter()
-            .map(|row| {
-                let distance: Option<f64> = row.get("distance");
-                let similarity = (1.0 - distance.unwrap_or(1.0)) as f32;
-                ChunkResult {
-                    document_id: row.get("document_id"),
-                    similarity_score: similarity,
-                    chunk_start_offset: row.get("chunk_start_offset"),
-                    chunk_end_offset: row.get("chunk_end_offset"),
-                    chunk_index: row.get("chunk_index"),
-                }
-            })
-            .collect();
-
-        Ok(chunk_results)
     }
 
     /// Find surrounding chunks for multiple center chunks from the same document with context window
